check_fraud/requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0

# Image processing
pillow>=10.0.0
opencv-python>=4.8.0

# Data handling
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# UI
streamlit>=1.29.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
numpy>=1.24.0
